EHR-2108 fix: Remove sorting on the front-end (#8)
* EHR-2108 fix: Remove sorting on the front-end

The backend already retrieves these lists sorted by rank and
description. Sorting on the front-end was resorting only by description
and removing the ranked sort.

* EHR-2108 fix: Add filter condition

Instead of removing conditional, add the filter condition to make sure
there are no regression on filtered fields. This still removes the sort
from the majority of the use cases and keeps the ranked sort.